requirements.txt
django[argon2]
django-colorfield
django-multiupload
django-storages
git+https://github.com/jazzband/django-polymorphic.git@v4.0.0a
git+https://git.private.coffee/kumi/django-multiselectfield.git
git+https://github.com/Chive/django-multiupload.git
dbsettings
django-filtersignals
mysqlclient
boto3
argon2_cffi
python-telegram-bot
python-dateutil
matrix-nio
holoviews
bokeh
panel
scipy
channels
django-autosecretkey
mozilla-django-oidc
python-gnupg
django-crispy-forms
crispy-bootstrap4
django-annoying